# swagger/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a Termy backend API.",
        "title": "Termy API",
        "contact": {
            "name": "API Support",
            "email": "andreysapozhkov535@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://mit-license.org/"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/elements": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "return all elements of structural models in the given text markup layer",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Elements"
                ],
                "summary": "Show all elements of structural models in the given text markup layer.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Text markup layer",
                        "name": "layer",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/git_iu7_bmstu_ru_sam20u745_testing-web-2023-labs_internal_domain_interfaces.OutputModelsDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "add new elements of structural models",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Elements"
                ],
                "summary": "Add new elements of structural models.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Text markup layer",
                        "name": "layer",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Structural model element names",
                        "name": "modelElementNames",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/git_iu7_bmstu_ru_sam20u745_testing-web-2023-labs_internal_domain_interfaces.ModelElementNamesDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/git_iu7_bmstu_ru_sam20u745_testing-web-2023-labs_internal_domain_interfaces.ModelElementsIdDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/layers": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "return all text markup layers",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Layers"
                ],
                "summary": "Show all text markup layers.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/git_iu7_bmstu_ru_sam20u745_testing-web-2023-labs_internal_domain_interfaces.LayersDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "add new text markup layer",
                "tags": [
                    "Layers"
                ],
                "summary": "Add new text markup layer.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Text markup layer",
                        "name": "layer",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/models": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "return all structural models in the given text markup layer",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Models"
                ],
                "summary": "Show all structural models in the given text markup layer.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Text markup layer",
                        "name": "layer",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/git_iu7_bmstu_ru_sam20u745_testing-web-2023-labs_internal_domain_interfaces.OutputModelsDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "add new structural models",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Models"
                ],
                "summary": "Add new structural models.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Text markup layer",
                        "name": "layer",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Structural model names",
                        "name": "modelNames",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/git_iu7_bmstu_ru_sam20u745_testing-web-2023-labs_internal_domain_interfaces.ModelNamesDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/git_iu7_bmstu_ru_sam20u745_testing-web-2023-labs_internal_domain_interfaces.ModelsIdDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/properties": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "return all unit properties",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Properties"
                ],
                "summary": "Show all unit properties.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/git_iu7_bmstu_ru_sam20u745_testing-web-2023-labs_internal_domain_interfaces.OutputModelsDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "add new unit properties",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Properties"
                ],
                "summary": "Add new unit properties.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Unit property names",
                        "name": "propertyNames",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/git_iu7_bmstu_ru_sam20u745_testing-web-2023-labs_internal_domain_interfaces.PropertyNamesDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/git_iu7_bmstu_ru_sam20u745_testing-web-2023-labs_internal_domain_interfaces.PropertiesIdDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/properties/unit": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "return all properties for the given unit in the given text markup layer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Properties"
                ],
                "summary": "Show all properties for the given unit in the given text markup layer.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Text markup layer",
                        "name": "layer",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Unit data according to which the search will be performed",
                        "name": "propertiesID",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/git_iu7_bmstu_ru_sam20u745_testing-web-2023-labs_internal_domain_interfaces.SearchUnitDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/git_iu7_bmstu_ru_sam20u745_testing-web-2023-labs_internal_domain_interfaces.OutputPropertiesDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "git_iu7_bmstu_ru_sam20u745_testing-web-2023-labs_internal_domain_interfaces.LayersDTO": {
            "type": "object",
            "properties": {
                "layers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "git_iu7_bmstu_ru_sam20u745_testing-web-2023-labs_internal_domain_interfaces.ModelElementNamesDTO": {
            "type": "object",
            "properties": {
                "model_element_names": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "git_iu7_bmstu_ru_sam20u745_testing-web-2023-labs_internal_domain_interfaces.ModelElementsIdDTO": {
            "type": "object",
            "properties": {
                "model_elements_id": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "git_iu7_bmstu_ru_sam20u745_testing-web-2023-labs_internal_domain_interfaces.ModelNamesDTO": {
            "type": "object",
            "properties": {
                "model_names": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "git_iu7_bmstu_ru_sam20u745_testing-web-2023-labs_internal_domain_interfaces.ModelsIdDTO": {
            "type": "object",
            "properties": {
                "models_id": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "git_iu7_bmstu_ru_sam20u745_testing-web-2023-labs_internal_domain_interfaces.OutputModelsDTO": {
            "type": "object",
            "properties": {
                "models": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/git_iu7_bmstu_ru_sam20u745_testing-web-2023-labs_internal_domain_interfaces.OutputModelDTO"
                    }
                }
            }
        },
        "git_iu7_bmstu_ru_sam20u745_testing-web-2023-labs_internal_domain_interfaces.OutputPropertiesDTO": {
            "type": "object",
            "properties": {
                "properties": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/git_iu7_bmstu_ru_sam20u745_testing-web-2023-labs_internal_domain_interfaces.OutputPropertyDTO"
                    }
                }
            }
        },
...
        "git_iu7_bmstu_ru_sam20u745_testing-web-2023-labs_internal_domain_interfaces.PropertiesIdDTO": {
            "type": "object",
            "properties": {
                "properties_id": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "git_iu7_bmstu_ru_sam20u745_testing-web-2023-labs_internal_domain_interfaces.PropertyNamesDTO": {
            "type": "object",
            "properties": {
                "property_names": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "git_iu7_bmstu_ru_sam20u745_testing-web-2023-labs_internal_domain_interfaces.SearchUnitDTO": {
            "type": "object",
            "properties": {
                "lang": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                }
            }
        },
...
    }
}